# docs/swagger.yaml
definitions:
...
  github_com_cossim_coss-server_internal_msg_api_http_model.BurnAfterReadingType:
    enum:
    - 0
    - 1
    type: integer
    x-enum-comments:
      IsBurnAfterReading: 阅后即焚消息
      NotBurnAfterReading: 非阅后即焚
    x-enum-varnames:
    - NotBurnAfterReading
    - IsBurnAfterReading
  github_com_cossim_coss-server_internal_msg_api_http_model.LabelMsgType:
    enum:
    - 0
    - 1
    type: integer
    x-enum-comments:
      IsLabel: 标注
      NotLabel: 不标注
    x-enum-varnames:
    - NotLabel
    - IsLabel
  github_com_cossim_coss-server_internal_msg_api_http_model.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  github_com_cossim_coss-server_internal_msg_api_http_model.SendGroupMsgRequest:
    properties:
      at_all_user:
        type: integer
      at_users:
        items:
          type: string
        type: array
      content:
        type: string
      dialog_id:
        type: integer
      group_id:
        type: integer
      is_burn_after_reading:
        $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.BurnAfterReadingType'
      replay_id:
        type: integer
      type:
        $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.UserMessageType'
    required:
    - content
    - dialog_id
    - group_id
    - type
    type: object
  github_com_cossim_coss-server_internal_msg_api_http_model.UserMessageType:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    - 9
    - 10
    - 11
    - 12
    type: integer
    x-enum-comments:
      MessageTypeCancelLabel: 取消标注
      MessageTypeDelete: 撤回消息
      MessageTypeEmojiReply: emoji回复
      MessageTypeFile: 文件消息
      MessageTypeImage: 图片消息
      MessageTypeLabel: 标注
      MessageTypeNotice: 群公告
      MessageTypeText: 文本消息
      MessageTypeVideo: 视频消息
      MessageTypeVideoCall: 视频通话
      MessageTypeVoice: 语音消息
      MessageTypeVoiceCall: 语音通话
    x-enum-varnames:
    - MessageTypeText
    - MessageTypeVoice
    - MessageTypeImage
    - MessageTypeLabel
    - MessageTypeNotice
    - MessageTypeFile
    - MessageTypeVideo
    - MessageTypeEmojiReply
    - MessageTypeVoiceCall
    - MessageTypeVideoCall
    - MessageTypeDelete
    - MessageTypeCancelLabel
...
  model.AfterMsg:
    properties:
      dialog_id:
        type: integer
      msg_id:
        type: integer
    type: object
...
  model.EditGroupMsgRequest:
    properties:
      content:
        description: New content of the message
        type: string
      msg_id:
        description: Message ID
        type: integer
      msg_type:
        type: integer
    required:
    - content
    - msg_id
    - msg_type
    type: object
  model.EditUserMsgRequest:
    properties:
      content:
        description: 消息内容
        type: string
      msg_id:
        description: Message ID
        type: integer
      msg_type:
        description: 消息类型
        type: integer
    required:
    - content
    - msg_id
    - msg_type
    type: object
...
  model.GetGroupMessageReadersResponse:
    properties:
      avatar:
        type: string
      name:
        type: string
      user_id:
        type: string
    type: object
  model.GroupMessageReadRequest:
    properties:
      dialog_id:
        description: 会话ID
        type: integer
      group_id:
        description: 群组ID
        type: integer
      msg_ids:
        description: 消息ID
        items:
          type: integer
        type: array
    required:
    - dialog_id
    - group_id
    - msg_ids
    type: object
...
  model.LabelGroupMessageRequest:
    properties:
      is_label:
        allOf:
        - $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.LabelMsgType'
        description: 是否标注
      msg_id:
        description: 消息ID
        type: integer
    required:
    - msg_id
    type: object
  model.LabelUserMessageRequest:
    properties:
      is_label:
        allOf:
        - $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.LabelMsgType'
        description: 是否标注
      msg_id:
        description: 消息ID
        type: integer
    required:
    - msg_id
    type: object
...
  model.ReadUserMsgsRequest:
    properties:
      dialog_id:
        description: 会话ID
        type: integer
      msg_ids:
        description: 消息id
        items:
          type: integer
        type: array
    required:
    - dialog_id
    - msg_ids
    type: object
  model.RecallGroupMsgRequest:
    properties:
      msg_id:
        description: Message ID
        type: integer
    required:
    - msg_id
    type: object
  model.RecallUserMsgRequest:
    properties:
      msg_id:
        description: Message ID
        type: integer
    required:
    - msg_id
    type: object
...
  model.SendUserMsgRequest:
    properties:
      content:
        type: string
      dialog_id:
        type: integer
      is_burn_after_reading:
        $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.BurnAfterReadingType'
      receiver_id:
        type: string
      replay_id:
        type: integer
      type:
        $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.UserMessageType'
    required:
    - content
    - dialog_id
    - receiver_id
    - type
    type: object
...
info:
  contact: {}
  title: CossApi
paths:
...
  /msg/after/get:
    post:
      consumes:
      - application/json
      description: 获取指定对话落后消息
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/model.AfterMsg'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.Response'
      summary: 获取指定对话落后消息
      tags:
      - Msg
...
  /msg/edit/group:
    post:
      consumes:
      - application/json
      description: 编辑群消息
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.EditGroupMsgRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.Response'
      summary: 编辑群消息
      tags:
      - Msg
  /msg/edit/user:
    post:
      consumes:
      - application/json
      description: 编辑用户消息
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.EditUserMsgRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.Response'
      summary: 编辑用户消息
      tags:
      - Msg
  /msg/group/read/get:
    get:
      consumes:
      - application/json
      description: 获取消息已读人员
      parameters:
      - description: 消息ID
        in: query
        name: msg_id
        required: true
        type: integer
      - description: 对话ID
        in: query
        name: dialog_id
        required: true
        type: integer
      - description: 群聊ID
        in: query
        name: group_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.GetGroupMessageReadersResponse'
                  type: array
              type: object
      summary: 获取消息已读人员
      tags:
      - Msg
  /msg/group/read/set:
    post:
      consumes:
      - application/json
      description: 批量设置群聊消息为已读
      parameters:
      - description: 请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.GroupMessageReadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.Response'
      summary: 批量设置群聊消息为已读
      tags:
      - Msg
  /msg/label/group:
    get:
      consumes:
      - application/json
      description: 获取群聊标注信息
      parameters:
      - description: 对话id
        in: query
        name: dialog_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.Response'
      summary: 获取群聊标注信息
      tags:
      - Msg
    post:
      consumes:
      - application/json
      description: 标注群聊消息状态
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LabelGroupMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.Response'
      summary: 标注群聊消息状态
      tags:
      - Msg
  /msg/label/user:
    get:
      consumes:
      - application/json
      description: 获取私聊标注信息
      parameters:
      - description: 对话id
        in: query
        name: dialog_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.Response'
      summary: 获取私聊标注信息
      tags:
      - Msg
    post:
      consumes:
      - application/json
      description: 标注用户消息状态
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LabelUserMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.Response'
      summary: 标注用户消息状态
      tags:
      - Msg
  /msg/list/group:
    get:
      consumes:
      - application/json
      description: 获取群聊消息
      parameters:
      - description: 群聊id
        in: query
        name: group_id
        required: true
        type: string
      - description: 用户id
        in: query
        name: user_id
        type: string
      - description: 类型
        in: query
        name: type
        type: string
      - description: 消息
        in: query
        name: content
        type: string
      - description: 页码
        in: query
        name: page_num
        required: true
        type: integer
      - description: 页大小
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.Response'
      summary: 获取群聊消息
      tags:
      - Msg
  /msg/list/user:
    get:
      consumes:
      - application/json
      description: 获取私聊消息
      parameters:
      - description: 用户id
        in: query
        name: user_id
        required: true
        type: string
      - description: 类型
        in: query
        name: type
        type: string
      - description: 消息
        in: query
        name: content
        type: string
      - description: 页码
        in: query
        name: page_num
        required: true
        type: integer
      - description: 页大小
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.Response'
      summary: 获取私聊消息
      tags:
      - Msg
  /msg/read/user:
    post:
      consumes:
      - application/json
      description: 批量设置私聊消息状态为已读
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ReadUserMsgsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.Response'
      summary: 批量设置私聊消息状态为已读
      tags:
      - Msg
  /msg/recall/group:
    post:
      consumes:
      - application/json
      description: 撤回群消息
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RecallGroupMsgRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.Response'
      summary: 撤回群消息
      tags:
      - Msg
  /msg/recall/user:
    post:
      consumes:
      - application/json
      description: 撤回用户消息
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RecallUserMsgRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.Response'
      summary: 撤回用户消息
      tags:
      - Msg
  /msg/send/group:
    post:
      consumes:
      - application/json
      description: 发送群聊消息
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.SendGroupMsgRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.Response'
      summary: 发送群聊消息
      tags:
      - Msg
  /msg/send/user:
    post:
      consumes:
      - application/json
      description: 发送私聊消息
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.SendUserMsgRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cossim_coss-server_internal_msg_api_http_model.Response'
      summary: 发送私聊消息
      tags:
      - Msg
...
swagger: "2.0"